+++ cmake/nfxGraphicsSources.cmake
@@ -10,6 +10,7 @@ set(NFX_GRAPHICS_SOURCES
     src/gl/pipeline/passes/DirectionalShadowPass.cpp
     src/gl/pipeline/passes/GeometryPass.cpp
     src/gl/pipeline/passes/GridPass.cpp
+    src/gl/pipeline/passes/PointShadowPass.cpp
     src/gl/pipeline/passes/PresentPass.cpp
     src/gl/pipeline/passes/SpotShadowPass.cpp
     src/gl/pipeline/passes/TransparentPass.cpp